move switcher to own file, add loading spinner
